chore: reorganizing the rtdn namespace
