+++ src/Domain/Rtdn/Type/RefundType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Imdhemy\GooglePlay\Domain\Rtdn;
+namespace Imdhemy\GooglePlay\Domain\Rtdn\Type;
 
 enum RefundType: int
 {
